loading css file separated storyListItem's layout fixed
