+++ src/App.css
@@ -6,10 +6,16 @@ a {
   color: inherit;
 }
 .app-container {
-  margin-top: 10px;
   width: 100%;
   max-width: 85%;
   margin: 10px auto 0px auto;
+}
+@media only screen and (max-width: 768px) {
+  .app-container {
+    width: 100%;
+    max-width: 100%;
+    margin: 0px auto 0px auto;
+  }
 }
 .clearfix::before,
 .clearfix::after {
@@ -28,7 +34,6 @@ a {
   border: 5px solid #333;
   border-radius: 30px;
   height: 30px;
-
   margin: 0 auto;
   opacity: 0;
   width: 30px;
@@ -48,7 +53,6 @@ a {
     opacity: 0;
   }
 }
-
 .menu {
   width: calc(100% - 10px);
   background-color: #ff6600;
@@ -59,6 +63,9 @@ a {
   font-weight: bold;
   vertical-align: middle;
   line-height: 26px;
+}
+.tab-menu-list {
+  padding: 0 8px 0 8px;
 }
 .tab-menu-list > ul {
   list-style-type: none;
@@ -71,6 +78,9 @@ a {
   border-right: 1px solid black;
   padding: 0px 8px 0px 8px;
   margin: 5px 0px 5px 0px;
+}
+.tab-menu-list > ul > li:first-child {
+  padding-left: 0;
 }
 .tab-menu-list > ul > li:last-child {
   border-right: 0;
@@ -141,7 +151,6 @@ a {
 .post-new > div {
   padding: 0 3px 0 3px;
 }
-
 /*SINGLE COMMENT*/
 .comment {
   margin-left: 2%;
