+++ src/loading.css
@@ -1,0 +1,23 @@
+.loading {
+  border: 5px solid #333;
+  border-radius: 30px;
+  height: 30px;
+  margin: 0 auto;
+  opacity: 0;
+  width: 30px;
+  animation: pulsate 1s ease-out;
+  animation-iteration-count: infinite;
+}
+@keyframes pulsate {
+  0% {
+    transform: scale(0.1);
+    opacity: 0;
+  }
+  50% {
+    opacity: 1;
+  }
+  100% {
+    transform: scale(1.2);
+    opacity: 0;
+  }
+}
